move util function
