+++ exposan/utils.py
@@ -14,6 +14,7 @@ for license details.
 
 import os, numpy as np, pandas as pd
 from chaospy import distributions as shape
+from thermosteam.functional import rho_to_V
 from qsdsan import ImpactItem, sanunits as su
 from qsdsan.utils import time_printer, AttrSetter
 from . import es_path
@@ -22,6 +23,7 @@ from . import es_path
 
 __all__ = (
     'add_fugitive_items',
+    'add_V_from_rho',
     'batch_setting_unit_params',
     'clear_unit_costs',
     'run_uncertainty',
@@ -32,6 +34,14 @@ def add_fugitive_items(unit, item_ID):
     unit._run()
     for i in unit.ins:
         i.stream_impact_item = ImpactItem.get_item(item_ID).copy(set_as_source=True)
+
+
+def add_V_from_rho(cmp, rho):
+    V_model = rho_to_V(rho, cmp.MW)
+    try: cmp.V.add_model(V_model)
+    except:
+        handle = getattr(cmp.V, cmp.locked_state)
+        handle.add_model(V_model)
 
 
 def batch_setting_unit_params(df, model, unit, exclude=()):
